requirements.txt
psycopg2==2.9.9
python-dotenv==1.0.1
SQLAlchemy[asyncio]==2.0.27
pgvector==0.2.5
SQLModel==0.0.16
asyncpg==0.29.0